requirements.txt
flask==3.1.2
flask-limiter==4.1.1
python-dotenv==1.2.1
PyPDF2==3.0.1
langchain==1.2.10
langchain-core==1.2.19
langchain-community==0.4.1
langchain-chroma==1.1.0
langchain-google-genai==4.2.1
langchain-text-splitters==1.1.1
langchain-huggingface
sentence-transformers==5.3.0
chromadb==1.5.5
google-genai==1.65.0
pydantic==2.12.5
gunicorn==25.1.0
